#2906: refactor: move auto feed update by interval helper, checkShouldRefreshFeeds(), to proper level of abstraction.
